+++ test/membrane/core/element/event_controller_test.exs
@@ -50,7 +50,6 @@ defmodule Membrane.Core.Element.EventControllerTest do
         playback: :playing,
         parent_pid: self(),
         synchronization: %{clock: nil, parent_clock: nil, stream_sync: nil},
-        handling_action?: false,
         delay_consuming_queues?: false,
         pads_to_snapshot: MapSet.new(),
         delayed_demands: MapSet.new(),