+++ test/membrane/core/element/lifecycle_controller_test.exs
@@ -49,7 +49,6 @@ defmodule Membrane.Core.Element.LifecycleControllerTest do
         playback: :playing,
         parent_pid: self(),
         synchronization: %{clock: nil, parent_clock: nil},
-        handling_action?: false,
         delay_consuming_queues?: false,
         pads_to_snapshot: MapSet.new(),
         delayed_demands: MapSet.new(),